minor updates to clients and list of detailed status fields and values from nov 5 2021 testing

+++ agents/acu/test_clients/point_130_45.py
@@ -12,4 +12,4 @@ def point_there(azimuth, elevation, waittime):
     acu_client.go_to.wait()
 
 if __name__ == "__main__":
-    point_there(float(130), float(45), 1)
+    point_there(float(130), float(25), 1)
